Add files via upload
Initial commit. So far, a basic page is set up that someone can choose a score. JS manipulates the DOM and the score can be store. Next step is to implement a backend to store a person's answer for the day. I also want to implement a calendar or a way for the app to keep track of the day.